backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
sqlmodel==0.0.22
pandas==2.2.2
scikit-learn==1.5.1
tensorflow-macos==2.15.0 
tensorflow-metal==1.1.0
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
